+++ frontend/packages/modelServing/src/extension-points/index.ts
@@ -1,11 +1,11 @@
 import { Extension } from '@openshift/dynamic-plugin-sdk';
 
 export type ModelServingPlatform = Extension<
-  'app.model-serving/platform',
+  'model-serving.platform',
   {
     id: string;
     name: string;
   }
 >;
 export const isModelServingPlatform = (extension: Extension): extension is ModelServingPlatform =>
-  extension.type === 'app.model-serving/platform';
+  extension.type === 'model-serving.platform';
