+++ frontend/src/k8sTypes.ts
@@ -1223,7 +1223,7 @@ export type DashboardCommonConfig = {
   disableNIMModelServing: boolean;
   disableAdminConnectionTypes: boolean;
   disableFineTuning: boolean;
-  disableModelServingExtension: boolean;
+  disableModelServingPlugin: boolean;
 };
 
 export type DashboardConfigKind = K8sResourceCommon & {
